feat: refactoring
layer, activation loss and statistic functions refactoring complete

+++ functions/statisticFunctions.cpp
@@ -5,7 +5,7 @@ using namespace std;
 float Mean(float *x, int size) {
   float sum = 0;
 
-  for (unsigned long int i = 0; i < size; i++)
+  for (int i = 0; i < size; i++)
     sum += x[i];
 
   return sum / size;
@@ -14,7 +14,7 @@ float Mean(float *x, int size) {
 float Max(float *x, int xSize) {
   float max = x[0];
 
-  for (unsigned long int i = 0; i < xSize; i++)
+  for (int i = 0; i < xSize; i++)
     if (x[i] > max)
       max = x[i];
 
@@ -24,7 +24,7 @@ float Max(float *x, int xSize) {
 float Min(float *x, int xSize) {
   float min = x[0];
 
-  for (unsigned long int i = 0; i < xSize; i++)
+  for (int i = 0; i < xSize; i++)
     if (x[i] < min)
       min = x[i];
 
@@ -35,7 +35,7 @@ void NormalizeN(std::vector<float> &input, std::vector<int> &minMax) {
   float min = input[0];
   float max = input[0];
 
-  for (unsigned long int i = 0; i < input.size(); i++) {
+  for (int i = 0; i < (int)input.size(); i++) {
     if (input[i] > max)
       max = input[i];
     if (input[i] < max)
@@ -45,7 +45,7 @@ void NormalizeN(std::vector<float> &input, std::vector<int> &minMax) {
   minMax[1] = max;
 
   float range = max - min;
-  for (int i = 0; i < input.size(); i++)
+  for (int i = 0; i < (int)input.size(); i++)
     input[i] = (input[i] - min) / range;
 }
 
@@ -54,14 +54,14 @@ void DeNormalizeN(std::vector<float> &input, std::vector<int> &minMax) {
   float max = minMax[1];
 
   float range = max - min;
-  for (int i = 0; i < input.size(); i++)
+  for (int i = 0; i < (int)input.size(); i++)
     input[i] = (int)(input[i] * range + min);
 }
 void Compress(float *input, int inputSize, int *minMax) {
   float min = input[0];
   float max = input[0];
 
-  for (unsigned long int i = 0; i < inputSize; i++) {
+  for (int i = 0; i < inputSize; i++) {
     if (input[i] >= max)
       max = input[i];
     if (input[i] <= max)
@@ -80,13 +80,13 @@ void DeCompress(std::vector<float> &input, std::vector<int> &minMax) {
   float max = minMax[1];
 
   float range = max - min;
-  for (int i = 0; i < input.size(); i++)
+  for (int i = 0; i < (int)input.size(); i++)
     input[i] = (int)(input[i] * range);
 }
 float VarianceOfPopulation(float *x, int size) {
   float mean = Mean(x, size), sumSq = 0;
 
-  for (unsigned long int i = 0; i < size; i++) {
+  for (int i = 0; i < size; i++) {
     float delta = x[i] - mean;
 
     sumSq += delta * delta;
@@ -103,7 +103,7 @@ void Standartize(float *dataset, int size) {
   auto mean = Mean(dataset, size);
   auto StDev = StandardDeviationOfPopulation(dataset, size);
 
-  for (unsigned long int i = 0; i < size; i++)
+  for (int i = 0; i < size; i++)
     dataset[i] = (dataset[i] - mean) / sqrt(StDev + Epsilon);
 }
 
@@ -114,7 +114,7 @@ void StandartizeLinearContract(float *dataset, int datasetSize, int *minMax,
   minMax[0] = min;
   minMax[1] = max;
   float range = max - min;
-  for (unsigned long int i = 0; i < datasetSize; i++) {
+  for (int i = 0; i < datasetSize; i++) {
     // if (dataset[i] != 0)
     dataset[i] = (end - start) * (dataset[i] - min) / range + start;
   }
@@ -125,7 +125,7 @@ void DeStandartizeLinearContract(float *dataset, int datasetSize, int *minMax) {
   float max = minMax[1];
   auto range = max - min;
 
-  for (unsigned long int i = 0; i < datasetSize; i++) {
+  for (int i = 0; i < datasetSize; i++) {
     // if (dataset[i] != 0)
     dataset[i] = (int)(range * (dataset[i] + 1.0) / 2.0 + min);
   }
@@ -134,7 +134,7 @@ void DeStandartizeLinearContract(float *dataset, int datasetSize, int *minMax) {
 void Normalize(float *dataset, int datasetSize) {
   const float Epsilon = 0.000001;
   float range = Max(dataset, datasetSize) - Min(dataset, datasetSize);
-  for (unsigned long int i = 0; i < datasetSize; i++) {
+  for (int i = 0; i < datasetSize; i++) {
     dataset[i] = dataset[i] / (range + Epsilon);
   }
 }
